refactory auth service

+++ src/app/services/database/projects.service.ts
@@ -3,20 +3,20 @@ import { SupabaseService } from '../supabase/supabase.service';
 import { Project } from '../../interfaces/project';
 import { ToastrService } from 'ngx-toastr'; // Importar ToastrService
 
-
 @Injectable({
   providedIn: 'root'
 })
 export class ProjectsService {
-
   private table = 'projects';
 
-  constructor(private supabaseService: SupabaseService,
-    private toastr: ToastrService ) {}
+  constructor(
+    private supabaseService: SupabaseService,
+    private toastr: ToastrService
+  ) {}
 
   async create(project: Partial<Project>): Promise<Project | null> {
     const session = await this.supabaseService.getSession();
-    const token = session?.data?.session?.access_token;
+    const token = session?.access_token;
 
     if (!token) {
       this.toastr.error('No hay token de sesión, el usuario no está autenticado.', 'Error');
@@ -26,8 +26,7 @@ export class ProjectsService {
     const { data, error } = await this.supabaseService.client
       .from(this.table)
       .insert(project)
-      .select()
-      ;
+      .select();
 
     if (error) {
       console.error('❌ Error al insertar proyecto en Supabase:', error);
@@ -46,13 +45,13 @@ export class ProjectsService {
 
   async getUserProjects(): Promise<Project[]> {
     const session = await this.supabaseService.getSession();
-    const token = session?.data?.session?.access_token;
+    const token = session?.access_token;
 
     if (!token) {
       throw new Error('❌ No hay token de sesión, el usuario no está autenticado.');
     }
 
-    const userId = session?.data?.session?.user?.id;
+    const userId = session?.user?.id;
 
     if (!userId) {
       throw new Error('❌ No se pudo obtener el ID del usuario.');
@@ -75,48 +74,53 @@ export class ProjectsService {
     const { data, error } = await this.supabaseService.client
       .from(this.table)
       .select('*')
-      .eq('githubusername', username)  
+      .eq('githubusername', username)
       .eq('name', projectName)
       .single();
-  
+
     if (error) {
       console.error('Error al obtener el proyecto:', error);
       return null;
     }
-  
+
     return data as Project;
   }
 
   async delete(id: number): Promise<boolean> {
     const session = await this.supabaseService.getSession();
-    const userId = session?.data?.session?.user?.id;
-  
+    const userId = session?.user?.id;
+
+    if (!userId) {
+      throw new Error('❌ No se pudo obtener el ID del usuario.');
+    }
+
     const { error } = await this.supabaseService.client
       .from(this.table)
       .delete()
       .eq('id', id);
-  
+
     if (error) {
       console.error('❌ Error al eliminar el proyecto:', error);
       throw error;
     }
-      return true;
+    return true;
   }
-  async getPaginated(page: number, pageSize: number = 24): Promise<{ data: Project[], total: number }> {
+
+  async getPaginated(page: number, pageSize: number = 24): Promise<{ data: Project[]; total: number }> {
     const from = (page - 1) * pageSize;
     const to = from + pageSize - 1;
-  
+
     const { data, error, count } = await this.supabaseService.client
       .from(this.table)
       .select('*', { count: 'exact' }) // Obtener total de registros
       .order('created_at', { ascending: false }) // Ordenar por fecha de creación
       .range(from, to); // Paginación
-  
+
     if (error) {
       console.error('❌ Error al obtener proyectos paginados:', error);
       throw error;
     }
-  
+
     return { data: data as Project[], total: count ?? 0 };
   }
